Add archived property to database object
